@@ -1,5 +1,5 @@
 import React from 'react'
-import Jumbotron from 'react-bootstrap/Jumbotron'
+import {Container} from 'react-bootstrap'
 import  RegistrationForm  from '../../components/registration-form/RegistrationForm.comp'
 
 import "./registration.style.css"
@@ -7,9 +7,9 @@ export const Registration = () => {
     return (
         <div className="registration-page bg-info">
             <div className="mt-5">
-            <Jumbotron className="form-box pb-3">
+            <Container className="form-box pb-3">
             <RegistrationForm/>
-            </Jumbotron>
+            </Container>
             </div>
         </div>
     )
